requirements.txt
Keras==2.3.1
matplotlib==3.1.3
numpy==1.19.2
opencv_python==3.4.1.15
pandas==1.1.5
Pillow==9.4.0
scikit_learn==1.2.1
tensorflow_gpu==1.13.1
